% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{coef.cv.tggl}
\alias{coef.cv.tggl}
\title{Extract coefficients from a cv.tggl object}
\usage{
\method{coef}{cv.tggl}(object, s = c("lambda.min"), ...)
}
\arguments{
\item{object}{A cv.tggl object.}

\item{s}{Value(s) of the penalty parameter lambda. Default is "lambda.min".}

\item{...}{Additional arguments passed to coef.tggl.}
}
\description{
Extract coefficients from a cv.tggl object
}
