% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{coef.tggl}
\alias{coef.tggl}
\title{Extract coefficients from a tggl object}
\usage{
\method{coef}{tggl}(object, s = NULL, exact = FALSE, ...)
}
\arguments{
\item{object}{A tggl object.}

\item{s}{Value(s) of the penalty parameter lambda at which coefficients are required.}

\item{exact}{Logical. If TRUE, forces exact calculation (not implemented yet, uses interpolation).}

\item{...}{Additional arguments.}
}
\description{
Extract coefficients from a tggl object
}
