% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{predict.cv.tggl}
\alias{predict.cv.tggl}
\title{Predict method for cv.tggl objects}
\usage{
\method{predict}{cv.tggl}(object, newx, s = c("lambda.min"), ...)
}
\description{
Predict method for cv.tggl objects
}
