% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{predict.tggl}
\alias{predict.tggl}
\title{Predict method for tggl objects}
\usage{
\method{predict}{tggl}(object, newx, s = NULL, ...)
}
\arguments{
\item{object}{A tggl object.}

\item{newx}{Matrix of new values for X.}

\item{s}{Value(s) of the penalty parameter lambda.}

\item{...}{Additional arguments.}
}
\description{
Predict method for tggl objects
}
